refactor: lib 폴더 정리
