+++ components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sanghee01/modal",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "A modal component for React",
   "type": "module",
   "types": "dist/index.d.ts",
