+++ hooks/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sanghee01/card-field-hooks",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "A set of hooks for managing card fields in React",
   "type": "module",
   "types": "dist/index.d.ts",
