+++ hooks/src/lib/index.ts
@@ -1,5 +1,5 @@
-export { useCardCompanySelect } from './useCardCompanySelect';
-export { useCardExpirationInput } from './useCardExpirationInput';
-export { useCardNumberInput } from './useCardNumberInput';
-export { useCardPasswordInput } from './useCardPasswordInput';
-export { useCvcInput } from './useCvcInput';
+export { useCardCompanySelect } from './hooks/useCardCompanySelect';
+export { useCardExpirationInput } from './hooks/useCardExpirationInput';
+export { useCardNumberInput } from './hooks/useCardNumberInput';
+export { useCardPasswordInput } from './hooks/useCardPasswordInput';
+export { useCvcInput } from './hooks/useCvcInput';
